+++ src/acc/utilities_impl.h
@@ -278,7 +278,7 @@ void makeNoiseImage(XFLOAT sigmaFudgeFactor,
 
 	// Copy the randomized image to A separate device-array, so that the
 	// transformer can be used to set up the actual particle image
-	for(int i=0; i<RandomImage.getSize(); i++)
+	for(size_t i=0; i<RandomImage.getSize(); i++)
 		RandomImage[i]=accMLO->transformer1.reals[i];
 
 #endif
@@ -296,7 +296,7 @@ static void TranslateAndNormCorrect(MultidimArray<RFLOAT > &img_in,
 	AccPtr<XFLOAT> temp = img_out.make<XFLOAT>(img_in.nzyxdim);
 	temp.allAlloc();
 
-	for (int i = 0; i < img_in.nzyxdim; i++)
+	for (unsigned long i = 0; i < img_in.nzyxdim; i++)
 		temp[i] = (XFLOAT) img_in.data[i];
 
 	temp.cpToDevice();
@@ -350,7 +350,7 @@ void normalizeAndTransformImage(	AccPtr<XFLOAT> &img_in,
 			accMLO->transformer1.forward();
 			accMLO->transformer1.fouriers.streamSync();
 
-			int FMultiBsize = ( (int) ceilf(( float)accMLO->transformer1.fouriers.getSize()*2/(float)BLOCK_SIZE));
+			size_t FMultiBsize = ( (int) ceilf(( float)accMLO->transformer1.fouriers.getSize()*2/(float)BLOCK_SIZE));
 			AccUtilities::multiply<XFLOAT>(FMultiBsize, BLOCK_SIZE, accMLO->transformer1.fouriers.getStream(),
 							(XFLOAT*)~accMLO->transformer1.fouriers,
 							(XFLOAT)1/((XFLOAT)(accMLO->transformer1.reals.getSize())),
@@ -371,7 +371,7 @@ void normalizeAndTransformImage(	AccPtr<XFLOAT> &img_in,
 			d_Fimg.cpToHost();
 			d_Fimg.streamSync();
 			img_out.initZeros(zSize, ySize, xSize);
-			for (int i = 0; i < img_out.nzyxdim; i ++)
+			for (unsigned long i = 0; i < img_out.nzyxdim; i ++)
 			{
 				img_out.data[i].real = (RFLOAT) d_Fimg[i].x;
 				img_out.data[i].imag = (RFLOAT) d_Fimg[i].y;
@@ -473,7 +473,7 @@ static void cosineFilter(
 void centerFFT_2D(int grid_size, int batch_size, int block_size,
 				cudaStream_t stream,
 				XFLOAT *img_in,
-				int image_size,
+				size_t image_size,
 				int xdim,
 				int ydim,
 				int xshift,
@@ -501,7 +501,7 @@ void centerFFT_2D(int grid_size, int batch_size, int block_size,
 
 void centerFFT_2D(int grid_size, int batch_size, int block_size,
 				XFLOAT *img_in,
-				int image_size,
+				size_t image_size,
 				int xdim,
 				int ydim,
 				int xshift,
@@ -568,8 +568,8 @@ void kernel_exponentiate_weights_fine(	XFLOAT *g_pdf_orientation,
 										bool *g_pdf_offset_zeros,
 										XFLOAT *g_weights,
 										XFLOAT min_diff2,
-										int oversamples_orient,
-										int oversamples_trans,
+										unsigned long  oversamples_orient,
+										unsigned long  oversamples_trans,
 										unsigned long *d_rot_id,
 										unsigned long *d_trans_idx,
 										unsigned long *d_job_idx,
