+++ src/ml_optimiser_mpi.cpp
@@ -820,7 +820,7 @@ void MlOptimiserMpi::expectation()
 			fn_tmp.compose("PPref_", i,"dat");
 			std::ofstream f;
 			f.open(fn_tmp.c_str());
-			for (unsigned j = 0; j < mymodel.PPref[i].data.nzyxdim; j++)
+			for (unsigned long j = 0; j < mymodel.PPref[i].data.nzyxdim; j++)
 					f << mymodel.PPref[i].data.data[j].real << std::endl;
 			f.close();
 		}
@@ -1048,11 +1048,11 @@ void MlOptimiserMpi::expectation()
 			int mdlX = mymodel.PPref[iclass].data.xdim;
 			int mdlY = mymodel.PPref[iclass].data.ydim;
 			int mdlZ = mymodel.PPref[iclass].data.zdim;
-			int mdlXYZ;
+			size_t mdlXYZ;
 			if(mdlZ == 0)
-				mdlXYZ = mdlX*mdlY;
+				mdlXYZ = (size_t)mdlX*(size_t)mdlY;
 			else
-				mdlXYZ = mdlX*mdlY*mdlZ;
+				mdlXYZ = (size_t)mdlX*(size_t)mdlY*(size_t)mdlZ;
 
 			try
 			{
@@ -1066,7 +1066,7 @@ void MlOptimiserMpi::expectation()
 			std::complex<XFLOAT> *pData = mdlClassComplex[iclass];
 
 			// Copy results into complex number array
-			for (unsigned long i = 0; i < mdlXYZ; i ++)
+			for (size_t i = 0; i < mdlXYZ; i ++)
 			{
 				std::complex<XFLOAT> arrayval(
 					(XFLOAT) mymodel.PPref[iclass].data.data[i].real,
